Move fractal to separate file
